64 bit compatibility in order to work with nexus files larger than 4GB

+++ src/common/nexusfile.h
@@ -12,12 +12,12 @@ namespace nx {
 		virtual ~NexusFile() {}
 		virtual void setFileName(const char* uri) = 0;
 		virtual bool open(OpenMode mode) = 0;
-		virtual int read(char* where, unsigned int length) = 0;
-		virtual int write(char* from, unsigned int length) = 0;
-		virtual int size() = 0;
-		virtual void* map(unsigned int from, unsigned int size) = 0;
+		virtual long long int read(char* where, size_t length) = 0;
+		virtual long long int write(char* from, size_t length) = 0;
+		virtual size_t size() = 0;
+		virtual void* map(size_t from, size_t size) = 0;
 		virtual bool unmap(void* mapped) = 0;
-		virtual bool seek(unsigned int to) = 0;
+		virtual bool seek(size_t to) = 0;
 	};
 }
